@@ -16,19 +16,24 @@
   "//": "DO NOT UPDATE @date-io/date-fns. It has some weird problems with nextjs",
   "dependencies": {
     "@date-io/date-fns": "1.3.13",
-    "@material-ui/core": "4.11.4",
-    "@material-ui/icons": "4.11.2",
-    "@material-ui/pickers": "3.3.10",
-    "@types/lodash": "4.14.178",
-    "@types/node": "16.0.0",
-    "@types/react": "17.0.38",
+    "@emotion/cache": "11.10.7",
+    "@emotion/react": "11.10.6",
+    "@emotion/server": "11.10.0",
+    "@emotion/styled": "11.10.6",
+    "@mui/icons-material": "5.11.16",
+    "@mui/material": "5.12.1",
+    "@mui/styles": "5.12.0",
+    "@mui/x-date-pickers": "6.2.0",
+    "@types/lodash": "4.14.194",
+    "@types/node": "16.18.23",
+    "@types/react": "18.0.37",
     "cross-env": "7.0.3",
     "date-fns": "2.28.0",
     "lodash": "4.17.21",
-    "next": "12.0.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "typescript": "4.5.5"
+    "next": "12.3.4",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "typescript": "5.0.4"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "5.19.0",
@@ -36,11 +41,11 @@
     "eslint": "8.13.0",
     "eslint-config-prettier": "8.5.0",
     "eslint-plugin-react": "7.29.4",
-    "eslint-plugin-react-hooks": "4.4.0",
-    "express": "4.17.2",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "express": "4.18.2",
     "http-proxy-middleware": "2.0.2",
-    "lint-staged": "12.3.1",
-    "prettier": "2.5.1"
+    "lint-staged": "13.2.1",
+    "prettier": "2.8.7"
   },
   "lint-staged": {
     "*.{js,json,md,css,scss}": [
