@@ -3,7 +3,7 @@ import getServerUrl from '../../../../util/serverUrl'
 import { PredictionCensored } from '../../../../shared'
 import Prediction from '../../../../components/prediction'
 import GoBackWrapper from '../../../../components/goBackWrapper'
-import { Button } from '@material-ui/core'
+import { Button } from '@mui/material'
 import { useState } from 'react'
 import { formatDateString } from '../../../../shared/date-util'
 import { removeUndefined } from '../../../../shared/object-util'
